requirements.txt
Django==5.0.3
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
Faker==33.0.0
ipython==8.29.0
Markdown==3.7
validate-docbr==1.10.0
